Replace generic object lambda in range agg with long primitive

+++ server/src/main/java/org/opensearch/search/optimization/filterrewrite/TreeTraversal.java
@@ -72,7 +72,7 @@ public final class TreeTraversal {
             this.activeIndex = packedValueRanges.firstRangeIndex(pointTree.getMinPackedValue(), pointTree.getMaxPackedValue());
         }
 
-        public int getActiveIndex() {
+        public long getActiveIndex() {
             return activeIndex;
         }
